+++ v1/server_websocket_test.go
@@ -15,11 +15,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestNewServerWebsocketGWS(t *testing.T) {
-	srv := testNewServer(t, apollows.WebsocketSubprotocolGraphqlWS, WithConnectTimeout(time.Second))
-
-	defer srv.Close()
-
+func testNewServerWebsocketGWS(t *testing.T, srv *httptest.Server) {
 	u := "ws" + strings.TrimPrefix(srv.URL, "http")
 
 	conn, resp, err := websocket.DefaultDialer.Dial(u, http.Header{
@@ -253,11 +249,7 @@ func TestNewServerWebsocketGWS(t *testing.T) {
 	assert.NoError(t, conn.Close())
 }
 
-func TestNewServerWebsocketGWTS(t *testing.T) {
-	srv := testNewServer(t, apollows.WebsocketSubprotocolGraphqlTransportWS, WithConnectTimeout(time.Second))
-
-	defer srv.Close()
-
+func testNewServerWebsocketGWTS(t *testing.T, srv *httptest.Server) {
 	u := "ws" + strings.TrimPrefix(srv.URL, "http")
 
 	conn, resp, err := websocket.DefaultDialer.Dial(u, http.Header{
@@ -478,6 +470,66 @@ func TestNewServerWebsocketGWTS(t *testing.T) {
 	assert.Equal(t, apollows.OperationComplete, msg.Type)
 
 	assert.NoError(t, conn.Close())
+}
+
+func TestNewServerWebsocketGWS(t *testing.T) {
+	srv := testNewServer(t, apollows.WebsocketSubprotocolGraphqlWS, WithConnectTimeout(time.Second))
+
+	defer srv.Close()
+
+	testNewServerWebsocketGWS(t, srv)
+}
+
+func TestNewServerWebsocketGWTS(t *testing.T) {
+	srv := testNewServer(t, apollows.WebsocketSubprotocolGraphqlTransportWS, WithConnectTimeout(time.Second))
+
+	defer srv.Close()
+
+	testNewServerWebsocketGWTS(t, srv)
+}
+
+func TestNewServerWebsocketGWSGWTS(t *testing.T) {
+	srv := testNewServerProtocols(
+		t,
+		[]apollows.Protocol{apollows.WebsocketSubprotocolGraphqlWS, apollows.WebsocketSubprotocolGraphqlTransportWS},
+		WithProtocol(apollows.WebsocketSubprotocolGraphqlTransportWS),
+		WithConnectTimeout(time.Second),
+	)
+
+	defer srv.Close()
+
+	testNewServerWebsocketGWS(t, srv)
+	testNewServerWebsocketGWTS(t, srv)
+}
+
+func TestNewServerWebsocketProtocolMismatch(t *testing.T) {
+	srv := testNewServerProtocols(
+		t,
+		[]apollows.Protocol{apollows.WebsocketSubprotocolGraphqlWS, apollows.WebsocketSubprotocolGraphqlTransportWS},
+		WithProtocol(apollows.WebsocketSubprotocolGraphqlTransportWS),
+		WithConnectTimeout(time.Second),
+	)
+
+	defer srv.Close()
+
+	u := "ws" + strings.TrimPrefix(srv.URL, "http")
+
+	conn, resp, err := websocket.DefaultDialer.Dial(u, http.Header{
+		"sec-websocket-protocol": []string{"foo"},
+	})
+
+	assert.NoError(t, err)
+
+	defer func() {
+		_ = conn.Close()
+		_ = resp.Body.Close()
+	}()
+
+	var msg apollows.Message
+
+	err = conn.ReadJSON(&msg)
+
+	assert.ErrorContains(t, err, apollows.ErrUnknownProtocol.Error())
 }
 
 func TestNewServerWebsocketKeepalive(t *testing.T) {
@@ -1255,7 +1307,7 @@ func TestNewServerWebsocketCombineErrorsGWS(t *testing.T) {
 	u := "ws" + strings.TrimPrefix(srv.URL, "http")
 
 	conn, resp, err := websocket.DefaultDialer.Dial(u, http.Header{
-		"sec-websocket-protocol": []string{string(apollows.WebsocketSubprotocolGraphqlTransportWS)},
+		"sec-websocket-protocol": []string{string(apollows.WebsocketSubprotocolGraphqlWS)},
 	})
 
 	assert.NoError(t, err)